Add data validation mechanism and endpoint to create airplane
Added validation mechanism in models, as I think the validation should
be placed there. The only thing that it is not validated it is the
foreign key constraint, so for example if we try to add an Airbus 777,
then an exception will be thrown, that's why we added we catch the
IntegrityError. Moreover, we added POST endpoint to create an airplane
entry.